created sidebar class
